+++ src/KingmakerGunslinger/ElementalRaces/ElementalRaceIdentityCatalog.cs
@@ -11,7 +11,7 @@ namespace KingmakerGunslinger.ElementalRaces
         internal const int HeritageIdentityCount = 53;
         internal const int FeatIdentityCount = 25;
         internal const int TraitFrameworkIdentityCount = 62;
-        internal const int TraitMechanicIdentityCount = 10;
+        internal const int TraitMechanicIdentityCount = 14;
         internal const int MechanicIdentityCount = LegacyMechanicIdentityCount +
             HeritageIdentityCount + FeatIdentityCount +
             TraitFrameworkIdentityCount + TraitMechanicIdentityCount;
@@ -231,7 +231,11 @@ namespace KingmakerGunslinger.ElementalRaces
                 "KMG.ElementalRaces.Traits.Ifrit.EfreetiMagic.ReducePerson",
                 "KMG.ElementalRaces.Traits.Oread.CrystallineForm.Resource",
                 "KMG.ElementalRaces.Traits.Oread.CrystallineForm.ArmedBuff",
-                "KMG.ElementalRaces.Traits.Oread.CrystallineForm.Mode"
+                "KMG.ElementalRaces.Traits.Oread.CrystallineForm.Mode",
+                "KMG.ElementalRaces.Traits.Undine.AcidBreath.Resource",
+                "KMG.ElementalRaces.Traits.Undine.AcidBreath.Ability",
+                "KMG.ElementalRaces.Traits.Undine.OozeBreath.Resource",
+                "KMG.ElementalRaces.Traits.Undine.OozeBreath.Ability"
             };
         }